Notion of order on the naturals and integers revised.
Now, both the integers and naturals use sr_precedes. So, in order to state
specifications of operations that can be implemented for both the integers
and naturals, the order sr_precedes can be used.

Also, sg_precedes, sr_precedes and field_precedes are now instances instead of
definitions. The priorities (respectively 10, 9 and 8) should ensure that the
right order should is selected. As a consequence we can use the notation ≤
everywhere.

Some theory with respect to the order on naturals (that does not hold for
arbitrary semirings) has been moved to orders/naturals.v. While it would be
convenient to move theory related to the order on integers to a separate file
too, that seems harder, because many non-order related theorems in
theory/integers.v, depend on theories about order.