Preliminary version including apartness and constructive order theory.
Before this commit a lot of our abstract theory on fields and order could not be applied to the constructive reals in CoRN. This was due to the fact that division was parametrized by a proof of (≠) (which does not contain constructive contents) and big parts of our abstract order theory required (≤) to be total or decidable.

From our previous experiences with CoRN we have learned that using CSetoids (setoids with apartness) everywhere is rather inconvenient in practice. Hence we only include apartness for fields instead of the whole hierarchy. Moreover, because setoid rewriting with Type'd relations does not work, we put apartness and strict order in Prop instead of Type. By incorporating type classes, and providing alternative constructors in case equality is decidable, one should not experience too much trouble from the apartness relation.

Some important changes:
* Some new classes, most importantly: StrongSetoid, StrongSetoid_Morphism, StrongInjective, Strong{Left,Right}Cancellation. Given instances of these classes we show that we have instances of Setoid, Setoid_Morphism, Injective, {Left,Right}Cancellation, respectively. One can go the other way around, in case equality is decidable, by hand.
* A complete revision of the order hierarchy. Instead of a total order, we now use a pseudo order. These interfaces are quite preliminary, especially for ordered rings and semirings: they could possibly be strengthened and tidied up.
* Include a class DecField, in which division is a total function (/0 = 0).
* A lot of renaming: Le instead of Order, Lt instead of StrictOrder, etc.
* Some tricks to avoid eager instance search (due to bug #2528).